+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * add new flatten method: AutoMask for flat surfaces with particles
 * `AFM.getCSV` creates a CSV file for export
+* fix setRange() argument in `plot.AFMdata`, so that pixels are not overwritten
 
 # nanoAFMr 2.3.2